# apps/backend/app/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Awade API",
    "description": "AI-powered educator support platform for African teachers",
    "version": "1.0.0"
  },
  "paths": {
    "/api/lesson-plans/generate": {
      "post": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Generate Lesson Plan",
        "description": "Generate a new lesson plan using AI.\nRequires educator authentication.",
        "operationId": "generate_lesson_plan_api_lesson_plans_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonPlanCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/lesson-plans/resources": {
      "get": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Get All Lesson Resources",
        "description": "Get all lesson resources for the current user.\nRequires authentication.",
        "operationId": "get_all_lesson_resources_api_lesson_plans_resources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LessonResourceResponse"
                  },
                  "type": "array",
                  "title": "Response Get All Lesson Resources Api Lesson Plans Resources Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/lesson-plans/resources/{resource_id}": {
      "get": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Get Lesson Resource",
        "description": "Get a specific lesson resource.\nRequires authentication.",
        "operationId": "get_lesson_resource_api_lesson_plans_resources__resource_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lesson-plans/": {
      "get": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Get Lesson Plans",
        "description": "Get lesson plans for the current user with optional filtering.\nRequires authentication.",
        "operationId": "get_lesson_plans_api_lesson_plans__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "subject",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subject"
            }
          },
          {
            "name": "grade_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grade Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonPlanResponse"
                  },
                  "title": "Response Get Lesson Plans Api Lesson Plans  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lesson-plans/{lesson_id}": {
      "get": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Get Lesson Plan",
        "description": "Get a specific lesson plan by ID.\nRequires authentication and ownership.",
        "operationId": "get_lesson_plan_api_lesson_plans__lesson_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Update Lesson Plan",
        "description": "Update a lesson plan.\nRequires educator or admin authentication and ownership.",
        "operationId": "update_lesson_plan_api_lesson_plans__lesson_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonPlanUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonPlanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Delete Lesson Plan",
        "description": "Delete a lesson plan.\nRequires educator or admin authentication and ownership.",
        "operationId": "delete_lesson_plan_api_lesson_plans__lesson_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lesson-plans/{lesson_id}/resources/generate": {
      "post": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Generate Lesson Resource",
        "description": "Generate AI-powered lesson resources for a specific lesson plan.\nRequires educator authentication.",
        "operationId": "generate_lesson_resource_api_lesson_plans__lesson_id__resources_generate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonResourceCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonResourceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/lesson-plans/{lesson_id}/resources": {
      "get": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Get Lesson Resources By Plan",
        "description": "Get all lesson resources for a specific lesson plan.\nRequires authentication and ownership.",
        "operationId": "get_lesson_resources_by_plan_api_lesson_plans__lesson_id__resources_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LessonResourceResponse"
                  },
                  "title": "Response Get Lesson Resources By Plan Api Lesson Plans  Lesson Id  Resources Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lesson-plans/resources/{resource_id}/export": {
      "post": {
        "tags": [
          "lesson-plans"
        ],
        "summary": "Export Lesson Resource",
        "description": "Export a lesson resource to PDF or DOCX format.\nRequires educator authentication.",
        "operationId": "export_lesson_resource_api_lesson_plans_resources__resource_id__export_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resource Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/contexts/lesson-plan/{lesson_plan_id}": {
      "get": {
        "tags": [
          "contexts"
        ],
        "summary": "Get Contexts By Lesson Plan",
        "description": "Get all contexts for a specific lesson plan.",
        "operationId": "get_contexts_by_lesson_plan_api_contexts_lesson_plan__lesson_plan_id__get",
        "parameters": [
          {
            "name": "lesson_plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lesson Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContextListResponse": {
        "properties": {
          "contexts": {
            "items": {
              "$ref": "#/components/schemas/ContextResponse"
            },
            "type": "array",
            "title": "Contexts"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "contexts",
          "total"
        ],
        "title": "ContextListResponse",
        "description": "Schema for list of contexts response."
      },
      "ContextResponse": {
        "properties": {
          "context_text": {
            "type": "string",
            "title": "Context Text",
            "description": "The context text content"
          },
          "context_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Type",
            "description": "Type of context (e.g., cultural, resources, student_background)"
          },
          "context_id": {
            "type": "integer",
            "title": "Context Id"
          },
          "lesson_plan_id": {
            "type": "integer",
            "title": "Lesson Plan Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "context_text",
          "context_id",
          "lesson_plan_id",
          "created_at",
          "updated_at"
        ],
        "title": "ContextResponse",
        "description": "Schema for context response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LessonPlanCreate": {
        "properties": {
          "subject": {
            "type": "string",
            "title": "Subject",
            "description": "Subject area (e.g., Mathematics, Science)"
          },
          "grade_level": {
            "type": "string",
            "title": "Grade Level",
            "description": "Grade level (e.g., Grade 4, Grade 7)"
          },
          "topic": {
            "type": "string",
            "title": "Topic",
            "description": "Specific topic within the subject (e.g., Fractions, Photosynthesis)"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id",
            "description": "User ID of the lesson plan author"
          }
        },
        "type": "object",
        "required": [
          "subject",
          "grade_level",
          "topic",
          "user_id"
        ],
        "title": "LessonPlanCreate",
        "description": "Schema for creating a new lesson plan."
      },
      "LessonPlanResponse": {
        "properties": {
          "lesson_id": {
            "type": "integer",
            "title": "Lesson Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "grade_level": {
            "type": "string",
            "title": "Grade Level"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "author_id": {
            "type": "integer",
            "title": "Author Id"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/LessonStatus"
          },
          "curriculum_learning_objectives": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curriculum Learning Objectives"
          },
          "curriculum_contents": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Curriculum Contents"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "title",
          "subject",
          "grade_level",
          "author_id",
          "created_at",
          "updated_at",
          "status"
        ],
        "title": "LessonPlanResponse",
        "description": "Schema for lesson plan response data."
      },
      "LessonPlanUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "grade_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade Level"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          },
          "duration_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Minutes"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LessonStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "learning_objectives": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Learning Objectives"
          },
          "topic_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Content"
          }
        },
        "type": "object",
        "title": "LessonPlanUpdate",
        "description": "Schema for updating an existing lesson plan."
      },
      "LessonResourceCreate": {
        "properties": {
          "lesson_plan_id": {
            "type": "integer",
            "title": "Lesson Plan Id"
          },
          "context_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Input"
          },
          "export_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Export Format"
          }
        },
        "type": "object",
        "required": [
          "lesson_plan_id"
        ],
        "title": "LessonResourceCreate",
        "description": "Schema for creating lesson resources."
      },
      "LessonResourceResponse": {
        "properties": {
          "lesson_resources_id": {
            "type": "integer",
            "title": "Lesson Resources Id"
          },
          "lesson_plan_id": {
            "type": "integer",
            "title": "Lesson Plan Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "context_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Input"
          },
          "ai_generated_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Generated Content"
          },
          "user_edited_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Edited Content"
          },
          "export_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Export Format"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "lesson_resources_id",
          "lesson_plan_id",
          "user_id",
          "status",
          "created_at"
        ],
        "title": "LessonResourceResponse",
        "description": "Schema for lesson resource response data."
      },
...
      "LessonStatus": {
        "type": "string",
        "enum": [
          "draft",
          "edited",
          "reviewed",
          "exported",
          "archived"
        ],
        "title": "LessonStatus",
        "description": "Enumeration of lesson plan status values."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}